:bug: Pasteboard preview list only shows supported types
